tests/requirements-test.txt
# Testing requirements
pytest>=7.0.0
pyyaml>=6.0
pandas>=1.5.0
numpy>=1.20.0
